# Optimize FileReadStream and BasicIStreamWrapper.