On (my) linux, perftest reports:
- ~40% gain for FileReadStream (`FileReadStream::Take()` loop),
- ~10% gain for ReaderParse_DummyHandler_FileReadStream.

With the same logic applied to BasicIStreamWrapper, which thus can now
also be created with a user buffer, performances align with those of
FileReadStream (see below, and as also noticed in #1365).

The "unbuffered" versions (added for FileReadStream) work solely with
the internal peekBuffer (Ch[4]) and are measured in perftest.  When
performances don't matter much, they can avoid the use of large
stack/heap buffers.

Though `appveyor` will probably have more trustworthy/relevant numbers, here are some from `perftest` on my linux laptop.

Codebase:
```
[ RUN      ] RapidJson.FileReadStream
[       OK ] RapidJson.FileReadStream (3292 ms)
[ RUN      ] RapidJson.FileReadStream_Unbuffered
[       OK ] RapidJson.FileReadStream_Unbuffered (6096 ms)
[ RUN      ] RapidJson.ReaderParse_DummyHandler_FileReadStream_SSE42
[       OK ] RapidJson.ReaderParse_DummyHandler_FileReadStream_SSE42 (8089 ms)
[ RUN      ] RapidJson.ReaderParse_DummyHandler_FileReadStream_Unbuffered_SSE42
[       OK ] RapidJson.ReaderParse_DummyHandler_FileReadStream_Unbuffered_SSE42 (10856 ms)
[ RUN      ] RapidJson.IStreamWrapper_Unbuffered
[       OK ] RapidJson.IStreamWrapper_Unbuffered (6383 ms)
[ RUN      ] RapidJson.IStreamWrapper_Setbuffered
[       OK ] RapidJson.IStreamWrapper_Setbuffered (6300 ms)
[ RUN      ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_Unbuffered_SSE42
[       OK ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_Unbuffered_SSE42 (16108 ms)
[ RUN      ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_Setbuffered_SSE42
[       OK ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_Setbuffered_SSE42 (15516 ms)
```

New code:
```
[ RUN      ] RapidJson.FileReadStream
[       OK ] RapidJson.FileReadStream (1889 ms)
[ RUN      ] RapidJson.FileReadStream_Unbuffered
[       OK ] RapidJson.FileReadStream_Unbuffered (4284 ms)
[ RUN      ] RapidJson.ReaderParse_DummyHandler_FileReadStream_SSE42
[       OK ] RapidJson.ReaderParse_DummyHandler_FileReadStream_SSE42 (7257 ms)
[ RUN      ] RapidJson.ReaderParse_DummyHandler_FileReadStream_Unbuffered_SSE42
[       OK ] RapidJson.ReaderParse_DummyHandler_FileReadStream_Unbuffered_SSE42 (10278 ms)
[ RUN      ] RapidJson.IStreamWrapper
[       OK ] RapidJson.IStreamWrapper (1871 ms)
[ RUN      ] RapidJson.IStreamWrapper_Unbuffered
[       OK ] RapidJson.IStreamWrapper_Unbuffered (4926 ms)
[ RUN      ] RapidJson.IStreamWrapper_Setbuffered
[       OK ] RapidJson.IStreamWrapper_Setbuffered (4920 ms)
[ RUN      ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_SSE42
[       OK ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_SSE42 (7266 ms)
[ RUN      ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_Unbuffered_SSE42
[       OK ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_Unbuffered_SSE42 (10844 ms)
[ RUN      ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_Setbuffered_SSE42
[       OK ] RapidJson.ReaderParse_DummyHandler_IStreamWrapper_Setbuffered_SSE42 (10557 ms)
```

The "_Setbuffered" tests show how it does not help (much) to set a custom/large `std::filebuf` to the `std::ifstream`. This is quite expected in the new code since reads are limited to `sizeof(peekBuffer_)` anyway, but even in codebase where `IStreamWrapper::Take()` calls `ifstream::get()` directly there is no/few improvement (the results for codebase come from simply/only adding this PR's "_SetBuffered" tests there).