@ylavic 2018-12-06
> Since it probably needs to be fixed regardless of this PR I'll propose a separate fix (another PR) for FileReadStream::Peek4().

See #1425